Applied reliability changes

+++ src/main/java/com/drajer/bsa/controller/NotificationContextController.java
@@ -53,11 +53,11 @@ public class NotificationContextController {
             notificationContextData.getId(), notificationContextData.getFhirServerBaseUrl(),
             notificationContextData.getNotificationResourceId(),
                 notificationContextData.getPatientId());
-    if (notificationContextDetails != null) {
+    if (notificationContextDetails != null && !notificationContextDetails.isEmpty()) {
       for (NotificationContext notificationContext : notificationContextDetails) {
         notificationContextService.delete(notificationContext);
-        return "NotificationContext deleted successfully.";
       }
+      return "NotificationContext deleted successfully.";
     }
     response.sendError(HttpServletResponse.SC_NOT_FOUND, "NotificationContext Not found");
     return "NotificationContext Not found";
